backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
motor==3.3.1
pydantic>=2.6.4
PyPDF2==3.0.1
python-docx==1.2.0
pytesseract==0.3.13
pillow==12.0.0
google-genai>=1.56.0
requests>=2.31.0